+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## [2026-01-22 14:58:02] - Version 0.1.0
+
+### 🚀 Changes:
+- 增加Accounts和Carriers纬度
+
+
 ## [2026-01-22 14:41:18] - Version 0.1.0
 
 ### 🚀 Changes: